Fix test coverage reports (#496)
