+++ package.json
@@ -11,13 +11,15 @@
     "mrproper": "npm run clean && lerna run mrproper",
     "postmrproper": "rm -rf node_modules cloudfront-dist/deployment/node_modules server/node_modules",
     "clean": "lerna exec -- rm -rf dist coverage deploy && lerna clean --yes",
-    "postclean": "rm -rf doc dist",
+    "postclean": "rm -rf doc dist coverage",
     "info": "lerna info",
     "prebuild": "rm -rf dist",
     "build": "lerna --stream --scope yourloops run build",
     "postbuild": "mkdir dist && mv packages/yourloops/dist dist/static",
     "gen-lambda": "node server/cloudfront-gen-lambda.js",
     "lint": "lerna run lint",
+    "clean-coverage": "rm -rf coverage",
+    "pretest": "npm run clean-coverage",
     "test": "npm run test-sundial && npm run test-tideline && npm run test-viz && npm run test-blip && npm run test-yourloops",
     "test-platform-client": "lerna --stream --scope=tidepool-platform-client run test",
     "test-sundial": "lerna --stream --scope=sundial run test",
@@ -97,7 +99,6 @@
     "karma": "^6.3.2",
     "karma-chai": "^0.1.0",
     "karma-chrome-launcher": "^3.1.0",
-    "karma-coverage": "^2.0.3",
     "karma-coverage-istanbul-reporter": "^3.0.3",
     "karma-firefox-launcher": "^2.1.0",
     "karma-mocha": "^2.0.1",
