+++ packages/yourloops/karma.conf.js
@@ -26,124 +26,16 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 /* eslint-disable @typescript-eslint/no-var-requires */
-const path = require("path");
-const _ = require("lodash");
+
+const { updateDefaultConfig, setKarmaConfig } = require("../../karma.common.conf");
 const webpack = require("./webpack.config.js");
 
-const isWSL = _.isString(process.env.WSL_DISTRO_NAME);
-const browsers = ["CustomChromeHeadless"];
-if (!isWSL) {
-  browsers.push("FirefoxHeadless");
-}
-
-delete webpack.entry;
-webpack.devtool = "inline-source-map";
-webpack.module.rules[0].options = { configFile: "tsconfig.test.json" };
-webpack.module.rules.push({
-  test: /\.tsx?$/,
-  exclude: /node_modules/,
-  loader: "@jsdevtools/coverage-istanbul-loader",
-  enforce: "post",
-  options: {
-    esModules: true,
-  },
-});
-
-const karmaConfig = {
-  autoWatch: false,
-  port: "8080",
-  browserNoActivityTimeout: 60000,
-  captureTimeout: 60000,
-  colors: true,
-  concurrency: 1,
-  singleRun: true,
-  browsers,
-  customLaunchers: {
-    CustomChromeHeadless: {
-      base: "ChromeHeadless",
-      flags: [
-        "--enable-automation",
-        "--no-default-browser-check",
-        "--no-first-run",
-        "--disable-default-apps",
-        "--disable-popup-blocking",
-        "--disable-translate",
-        "--disable-background-timer-throttling",
-        "--disable-renderer-backgrounding",
-        "--disable-device-discovery-notifications",
-        "--disable-dev-shm-usage",
-        "--disable-gpu",
-        "--headless",
-        "--no-sandbox",
-        "--remote-debugging-port=9222",
-      ],
-    },
-  },
-  coverageIstanbulReporter: {
-    // reports can be any that are listed here: https://github.com/istanbuljs/istanbuljs/tree/73c25ce79f91010d1ff073aa6ff3fd01114f90db/packages/istanbul-reports/lib
-    reports: ["html", "text-summary"],
-    // base output directory. If you include %browser% in the path it will be replaced with the karma browser name
-    dir: path.join(__dirname, "coverage"),
-    // Combines coverage information from multiple browsers into one report rather than outputting a report for each browser.
-    combineBrowserReports: true,
-    // if using webpack and pre-loaders, work around webpack breaking the source path
-    fixWebpackSourcePaths: false,
-    // Omit files with no statements, no functions and no branches covered from the report
-    skipFilesWithNoCoverage: true,
-    // Most reporters accept additional config options. You can pass these through the `report-config` option
-    "report-config": {
-      // all options available at: https://github.com/istanbuljs/istanbuljs/blob/73c25ce79f91010d1ff073aa6ff3fd01114f90db/packages/istanbul-reports/lib/html/index.js#L257-L261
-      html: {
-        // outputs the report in ./coverage/html
-        subdir: "html",
-      },
-    },
-    // enforce percentage thresholds
-    // anything under these percentages will cause karma to fail with an exit code of 1 if not running in watch mode
-    thresholds: {
-      emitWarning: true, // set to `true` to not fail the test command when thresholds are not met
-      // thresholds for all files
-      global: {
-        statements: 90,
-        lines: 95,
-        branches: 95,
-        functions: 100,
-      },
-      // thresholds per file
-      // each: {
-      //   statements: 100,
-      //   lines: 100,
-      //   branches: 100,
-      //   functions: 100,
-      //   overrides: {
-      //     'baz/component/**/*.js': {
-      //       statements: 98
-      //     }
-      //   }
-      // },
-    },
-
-    verbose: true, // output config used by istanbul for debugging
-  },
-  mime: {
-    "text/x-typescript": ["ts", "tsx"],
-  },
+const config = {
   files: ["test/index.ts"],
-  frameworks: ["mocha"],
   preprocessors: {
-    "test/index.ts": "webpack",
-  },
-  reporters: ["mocha", "coverage-istanbul"],
-  webpack,
-  webpackMiddleware: {
-    noInfo: true,
-    stats: "errors-only",
+    "test/index.ts": ["webpack", "sourcemap"],
   },
 };
-
-function setKarmaConfig(config) {
-  karmaConfig.logLevel = config.LOG_INFO;
-  config.set(karmaConfig);
-}
+updateDefaultConfig("yourloops", config, webpack, true);
 
 module.exports = setKarmaConfig;
