+++ packages/yourloops/package.json
@@ -14,7 +14,7 @@
   "license": "BSD-2-Clause",
   "scripts": {
     "mrproper": "rm -rf node_modules dist",
-    "clean": "rm -rf dist coverage",
+    "clean": "rm -rf dist",
     "lint": "eslint app components lib models pages test types",
     "build": "npm run build-prod",
     "prebuild-prod": "npm run clean",
@@ -22,7 +22,6 @@
     "prebuild-dev": "npm run clean",
     "build-dev": "NODE_ENV=development webpack --mode=development",
     "start-dev": "NODE_ENV=development NODE_OPTIONS='--max-old-space-size=4096' USE_WEBPACK_DEV_SERVER=true webpack-dev-server --cache --colors --port 3001 --host 0.0.0.0",
-    "pretest": "rm -rf coverage",
     "test": "NODE_ENV='test' NODE_OPTIONS='--max-old-space-size=4096' karma --single-run --log-level debug start"
   },
   "dependencies": {
